Allow int32 featureIds, NodeType similar to SIMPL, FaceLabels use 0
* Int32 for FeatureIds as the type
    * Still creating an entire copy of the FeatureId array
* If a node is part of a triangle that has a ZERO label, then add 10 to the node type
* SurfaceNets uses int32::max() for specific values, this is changed after the algorithm is complete to be ZERO.

+++ src/Plugins/ComplexCore/src/ComplexCore/SurfaceNets/MMGeometryGL.cpp
@@ -74,7 +74,7 @@ MMGeometryGL::MMGeometryGL(MMSurfaceNet* surfaceNet)
   for(int idxVtx = 0; idxVtx < cellMap->numVertices(); idxVtx++)
   {
     float vertexPositions[12];
-    unsigned short labelsTmp[2];
+    int32_t labelsTmp[2];
 
     // Back-bottom edge
     if(cellMap->getEdgeQuad(idxVtx, MMCellFlag::Edge::BackBottomEdge, vertexPositions, labelsTmp) == true)
@@ -127,7 +127,7 @@ void MMGeometryGL::maxSize(float size[3])
   size[2] = m_size[2];
 }
 
-void MMGeometryGL::makeGLQuad(float* positions, unsigned short tissueLabels[2], float* quadVerts, unsigned int* quadIndices, int idxOffset)
+void MMGeometryGL::makeGLQuad(float* positions, int32_t tissueLabels[2], float* quadVerts, unsigned int* quadIndices, int idxOffset)
 {
   float norm[3];
   computeQuadNormal(positions, norm);
